removed metadata from layout
