+++ app/dashboard/migration/page.tsx
@@ -1,44 +1,11 @@
 "use client";
 
-import type { Metadata } from "next";
 import Questions from "@/components/widgets/questions";
 import StakingDetails from "@/components/widgets/staking-details";
 import { useNetwork } from "wagmi";
 import { Button } from "@/components/ui/button";
 import MigrationDetails from "@/components/widgets/migration-details";
-
-const metadata: Metadata = {
-  title: "Proof of X | Migration center",
-  description:
-    "During the migration period, you can deposit $EULER to get the new token and claim your membership NFTs.",
-  openGraph: {
-    title: "Proof of X | Migration center",
-    description:
-      "During the migration period, you can deposit $EULER to get the new token and claim your membership NFTs.",
-
-    url: "https://pox.me/dashboard/migration",
-    siteName: "Proof of X",
-    images: [
-      {
-        url: "/images/home-hero.jpg",
-        width: 1800,
-        height: 1600,
-        alt: "A cool fingerprint image",
-      },
-    ],
-    locale: "en_US",
-    type: "website",
-  },
-  twitter: {
-    card: "summary_large_image",
-    title: "Proof of X | Migration center",
-    description:
-      "During the migration period, you can deposit $EULER to get the new token and claim your membership NFTs.",
-
-    creator: "@proofxme",
-    images: ["/images/home-hero.jpg"],
-  },
-};
+import { metadata } from "./metadata";
 
 export default function Home() {
   const { chain } = useNetwork();
@@ -114,3 +81,5 @@ export default function Home() {
     </main>
   );
 }
+
+Home.metadata = metadata;
