+++ app/layout.tsx
@@ -7,34 +7,6 @@ import { Analytics } from "@vercel/analytics/react";
 
 const inter = Inter({ subsets: ["latin"] });
 
-export const metadata: Metadata = {
-  title: "Proof of X | Home",
-  description: "Proof of X | Home",
-  openGraph: {
-    title: "Proof of X | Home",
-    description: "Proof of X | Home",
-    url: "https://pox.me",
-    siteName: "Proof of X",
-    images: [
-      {
-        url: "/images/home-hero.jpg",
-        width: 1800,
-        height: 1600,
-        alt: "A cool fingerprint image",
-      },
-    ],
-    locale: "en_US",
-    type: "website",
-  },
-  twitter: {
-    card: "summary_large_image",
-    title: "Proof of X | Home",
-    description: "Proof of X | Home",
-    creator: "@proofxme",
-    images: ["/images/home-hero.jpg"],
-  },
-};
-
 export default function RootLayout({
   children,
 }: {
