+++ app/membership/metadata.ts
@@ -1,0 +1,33 @@
+import type { Metadata } from "next";
+export const metadata: Metadata = {
+  title: "Proof of X | Membership",
+  description:
+    "Get unlimited access to all our premium features with a lifetime membership.",
+  openGraph: {
+    title: "Proof of X | Membership",
+    description:
+      "Get unlimited access to all our premium features with a lifetime membership.",
+
+    url: "https://pox.me/membership",
+    siteName: "Proof of X",
+    images: [
+      {
+        url: "https://assets.pox.me/memberships/membership.png",
+        width: 1800,
+        height: 1600,
+        alt: "Proof of X lifetime membership image",
+      },
+    ],
+    locale: "en_US",
+    type: "website",
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "Proof of X | Membership",
+    description:
+      "Get unlimited access to all our premium features with a lifetime membership.",
+
+    creator: "@proofxme",
+    images: ["https://assets.pox.me/memberships/membership.png"],
+  },
+};
