+++ app/roadmap/page.tsx
@@ -1,43 +1,10 @@
 "use client";
 
 import { JSX, SVGProps } from "react";
-import type { Metadata } from "next";
-
-const metadata: Metadata = {
-  title: "Proof of X | Roadmap",
-  description:
-    "We're excited to share what we're working on. Here are the features and improvements you can expect in the coming months.",
-  openGraph: {
-    title: "Proof of X | Roadmap",
-    description:
-      "We're excited to share what we're working on. Here are the features and improvements you can expect in the coming months.",
-
-    url: "https://pox.me/roadmap",
-    siteName: "Proof of X",
-    images: [
-      {
-        url: "/images/home-hero.jpg",
-        width: 1800,
-        height: 1600,
-        alt: "A cool fingerprint image",
-      },
-    ],
-    locale: "en_US",
-    type: "website",
-  },
-  twitter: {
-    card: "summary_large_image",
-    title: "Proof of X | Roadmap",
-    description:
-      "We're excited to share what we're working on. Here are the features and improvements you can expect in the coming months.",
-
-    creator: "@proofxme",
-    images: ["/images/home-hero.jpg"],
-  },
-};
 
 // Assuming the JSON data is imported like this
 import roadmapData from "./roadmapData.json"; // replace './roadmapData.json' with the actual path to your JSON file
+import { metadata } from "./metadata";
 
 export default function Component() {
   return (
@@ -100,6 +67,8 @@ export default function Component() {
     </section>
   );
 }
+
+Component.metadata = metadata;
 
 function CheckIcon(props: JSX.IntrinsicAttributes & SVGProps<SVGSVGElement>) {
   return (
